Rename frigate package to events
